Move nodes providers to separate package
