--- sonar-tsql-plugin/src/main/java/org/sonar/plugins/tsql/sensors/custom/INodesProvider.java
@@ -1,5 +1,0 @@
-package org.sonar.plugins.tsql.sensors.custom;
-
-public interface INodesProvider {
-	ParsedNode[] getNodes(final ParsedNode node);
-}
