+++ sonar-tsql-plugin/src/main/java/org/sonar/plugins/tsql/sensors/custom/providers/ChildrenNodesProvider.java
@@ -1,9 +1,10 @@
-package org.sonar.plugins.tsql.sensors.custom;
+package org.sonar.plugins.tsql.sensors.custom.providers;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import org.antlr.v4.runtime.tree.ParseTree;
+import org.sonar.plugins.tsql.sensors.custom.ParsedNode;
 
 public class ChildrenNodesProvider implements INodesProvider {
 
@@ -15,11 +16,11 @@ public class ChildrenNodesProvider implements INodesProvider {
 	}
 
 	void visit(List<ParsedNode> nodes, final ParsedNode root) {
-		if (root  == null || root.getItem() == null) {
+		if (root == null || root.getItem() == null) {
 			return;
 		}
 		final ParseTree tree = root.getItem();
-		
+
 		final int c = tree.getChildCount();
 
 		for (int i = 0; i < c; i++) {
