+++ sonar-tsql-plugin/src/main/java/org/sonar/plugins/tsql/sensors/custom/providers/INodesProvider.java
@@ -1,0 +1,7 @@
+package org.sonar.plugins.tsql.sensors.custom.providers;
+
+import org.sonar.plugins.tsql.sensors.custom.ParsedNode;
+
+public interface INodesProvider {
+	ParsedNode[] getNodes(final ParsedNode node);
+}
