+++ sonar-tsql-plugin/src/main/java/org/sonar/plugins/tsql/sensors/custom/providers/SiblingsNodesProvider.java
@@ -1,13 +1,13 @@
-package org.sonar.plugins.tsql.sensors.custom;
+package org.sonar.plugins.tsql.sensors.custom.providers;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import org.antlr.v4.runtime.tree.ParseTree;
+import org.sonar.plugins.tsql.sensors.custom.ParsedNode;
 
 public class SiblingsNodesProvider implements INodesProvider {
 
-	@Override
 	public ParsedNode[] getNodes(final ParsedNode node) {
 		final List<ParsedNode> nodes = new ArrayList<ParsedNode>();
