+++ sonar-tsql-plugin/src/test/java/org/sonar/plugins/tsql/sensors/custom/providers/SiblingNodesProviderTest.java
@@ -1,16 +1,17 @@
-package org.sonar.plugins.tsql.sensors.custom;
+package org.sonar.plugins.tsql.sensors.custom.providers;
 
 import org.antlr.v4.runtime.tree.ParseTree;
 import org.junit.Assert;
 import org.junit.Test;
 import org.sonar.plugins.tsql.helpers.Antlr4Utils;
 import org.sonar.plugins.tsql.rules.custom.Rule;
+import org.sonar.plugins.tsql.sensors.custom.ParsedNode;
 
-public class ChildrenNodesProviderTest {
+public class SiblingNodesProviderTest {
 
 	@Test
 	public void test() {
-		ChildrenNodesProvider provider = new ChildrenNodesProvider();
+		SiblingsNodesProvider provider = new SiblingsNodesProvider();
 		Assert.assertEquals(0, provider.getNodes(null).length);
 	}
 
@@ -18,9 +19,9 @@ public class ChildrenNodesProviderTest {
 	public void test2() {
 		ParseTree tree = Antlr4Utils.get("select * from dbo.test");
 		Rule rule = new Rule();
-		ChildrenNodesProvider provider = new ChildrenNodesProvider();
-		ParsedNode node = new ParsedNode(tree, rule, "test");
-		Assert.assertEquals(26, provider.getNodes(node).length);
+		SiblingsNodesProvider provider = new SiblingsNodesProvider();
+		ParsedNode node = new ParsedNode(tree.getChild(0).getChild(0), rule, "test");
+		Assert.assertEquals(1, provider.getNodes(node).length);
 	}
 
 }
